requirements.txt
numpy~=2.2.6
pillow~=11.3.0
torch~=2.8.0
albumentations~=2.0.8
tqdm~=4.67.1
torchvision~=0.23.0
